chore: splitting the app into more components
