+++ src/components/Header/index.tsx
@@ -1,0 +1,28 @@
+import { Flex } from "@chakra-ui/react";
+import { Logo } from "./Logo";
+import { NotificationsNav } from "./NotificationsNav";
+import { Profile } from "./Profile";
+import { Search } from "./Search";
+
+export function Header() {
+  return (
+    <Flex
+      w="100%"
+      as="header"
+      maxWidth={1480}
+      h="20"
+      mx="auto"
+      mt="4"
+      align="center"
+      px="6"
+    >
+      <Logo />
+      <Search />
+
+      <Flex align="center" ml="auto">
+        <NotificationsNav />
+        <Profile />
+      </Flex>
+    </Flex>
+  );
+}
